% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-data.R
\docType{data}
\name{barney_schaffner_18}
\alias{barney_schaffner_18}
\title{Dataset from Barney and Schaffner (2018)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8512 rows and 6 columns.}
\usage{
barney_schaffner_18
}
\description{
Dataset from Barney and Schaffner (2018)
}
\references{
Barney and Schaffner (2018)
}
\keyword{dataset}
