% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-functions.R
\name{block_sample}
\alias{block_sample}
\title{Block re-sampling for block bootstrap}
\usage{
block_sample(dat, id_cluster, id_unique, J, max_cluster_size)
}
\arguments{
\item{dat}{a data frame with \code{Ynew}, \code{Yold} and \code{treat}.}

\item{id_cluster}{a cluster id vector of length n.}
}
\value{
a list of resampled data.
}
\description{
Block re-sampling for block bootstrap
}
\keyword{internal}
