% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-probit.R
\name{count_nj}
\alias{count_nj}
\title{Counting category selections by groups}
\usage{
count_nj(Y, id_group, n_group, n_cat, j_max, j_min)
}
\description{
Counting category selections by groups
}
\keyword{internal}
