% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-helper.R
\name{create_group_dummy}
\alias{create_group_dummy}
\title{Create a group indicator
The same-length condition is already checked by \code{ord_did_check_input}.}
\usage{
create_group_dummy(Ynew, Yold, treat, pre = FALSE)
}
\description{
Create a group indicator
The same-length condition is already checked by \code{ord_did_check_input}.
}
\keyword{internal}
