% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dat_block_boot}
\alias{dat_block_boot}
\title{helper function for the loss computation}
\usage{
dat_block_boot(dat, id_cluster, id_cluster_boot, max_cluster_size)
}
\description{
helper function for the loss computation
}
\keyword{internal}
