% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-semiparametric-step1.r
\name{.compress_ties}
\alias{.compress_ties}
\title{Compress tied observations}
\usage{
.compress_ties(u_sorted, y_sorted, w_sorted, tol = 1e-12)
}
\description{
Compress tied observations
}
\keyword{internal}
