% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-semiparametric-npmle.R
\name{.compute_joint_npmle}
\alias{.compute_joint_npmle}
\title{Compute joint NPMLE for error distribution using all ordinal categories}
\usage{
.compute_joint_npmle(Y00, X00, b_00, alpha)
}
\arguments{
\item{Y00}{outcomes for (D=0, t=0) cell}

\item{X00}{design matrix}

\item{b_00}{Step 1 index coefficients}

\item{alpha}{cutoff parameter}
}
\value{
list with Ftilde, knots, values; or NULL on failure
}
\description{
Treats ordered response as interval-censored data:
  Y = min: epsilon in (-Inf, u_i]
  Y = middle: epsilon in (u_i, u_i+alpha]
  Y = max: epsilon in (u_i+alpha, Inf)
}
\keyword{internal}
