% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-semiparametric-npmle.R
\name{.estimate_alpha_smooth}
\alias{.estimate_alpha_smooth}
\title{Estimate cutoff alpha using the smoothed Ftilde}
\usage{
.estimate_alpha_smooth(Y00, X00, b_00, Ftilde)
}
\arguments{
\item{Y00}{outcomes for (D=0, t=0) cell}

\item{X00}{design matrix}

\item{b_00}{Step 1 index coefficients}

\item{Ftilde}{smoothed CDF function}
}
\value{
estimated alpha, or NA if root-finding fails
}
\description{
Solves: mean(1 - 1\{Y = max(Y)\} - Ftilde(X'b + alpha)) = 0
}
\keyword{internal}
