% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid.R
\name{.extract_semiparametric_cells}
\alias{.extract_semiparametric_cells}
\title{Extract cell data for semiparametric estimation}
\usage{
.extract_semiparametric_cells(data, outcome, post, treat, x_cols)
}
\description{
Extract cell data for semiparametric estimation
}
\keyword{internal}
