% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-semiparametric-step1.r
\name{.init_beta_probit}
\alias{.init_beta_probit}
\title{Initialize beta via probit regression}
\usage{
.init_beta_probit(Y, X, fixed_index)
}
\description{
Initialize beta via probit regression
}
\keyword{internal}
