% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-probit.R
\name{.LogLikelihoodProbit}
\alias{.LogLikelihoodProbit}
\title{Log-likelihood Function for ordered Probit}
\usage{
.LogLikelihoodProbit(par, Y, cutoff_val = NULL)
}
\description{
Log-likelihood Function for ordered Probit
}
\keyword{internal}
