% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-semiparametric-step1.r
\name{.pava_weighted}
\alias{.pava_weighted}
\title{Weighted Pool Adjacent Violators Algorithm (PAVA)}
\usage{
.pava_weighted(mu, wt)
}
\description{
Weighted Pool Adjacent Violators Algorithm (PAVA)
}
\keyword{internal}
