% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-semiparametric-smoothing.R
\name{.smooth_Fhat}
\alias{.smooth_Fhat}
\title{Smooth an isotonic step-function CDF using I-splines}
\usage{
.smooth_Fhat(Fhat, n0, K = NULL, degree = 3)
}
\arguments{
\item{Fhat}{FhatIsotonic object (or list with $knots, $values)}

\item{n0}{sample size of the (D=0,t=0) cell (for choosing K)}

\item{K}{number of internal knots; if NULL, uses floor(n0^(1/3))}

\item{degree}{spline degree (default 3, giving C^2 smoothness)}
}
\value{
list with Ftilde (smoothed CDF), ftilde (density), coef, knot_seq
}
\description{
Smooth an isotonic step-function CDF using I-splines
}
\keyword{internal}
