% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-semiparametric-step2.r
\name{.solve_BB_locscale}
\alias{.solve_BB_locscale}
\title{BB::dfsane solver wrapper for location-scale model}
\usage{
.solve_BB_locscale(moment_map, theta_init, maxit, tol, verbose)
}
\description{
BB::dfsane solver wrapper for location-scale model
}
\keyword{internal}
