% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-semiparametric-step1.r
\name{.solve_spectral}
\alias{.solve_spectral}
\title{Spectral method (Barzilai-Borwein with backtracking)}
\usage{
.solve_spectral(
  moment_map,
  b_init,
  maxit,
  tol,
  bt_factor,
  rel_improve,
  verbose
)
}
\description{
Spectral method (Barzilai-Borwein with backtracking)
}
\keyword{internal}
