% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-staggered-internals.R
\name{.staggered_aggregate}
\alias{.staggered_aggregate}
\title{Aggregate group-time effects using weights}
\usage{
.staggered_aggregate(gt_results, weights = "uniform")
}
\arguments{
\item{gt_results}{list of group-time results from .staggered_estimate_all}

\item{weights}{"uniform" or named list}
}
\value{
list with aggregate zeta and tau
}
\description{
Aggregate group-time effects using weights
}
\keyword{internal}
