% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-diagnostic.R
\name{equivalence_test}
\alias{equivalence_test}
\title{Conduct an equivalence test of q1(v) = q0(v)}
\usage{
equivalence_test(theta, vcov, alpha, epsilon)
}
\arguments{
\item{theta}{a list of parameters}

\item{vcov}{a variance covariance matrix of theta}

\item{alpha}{level of a test. Should be between 0 and 1.}

\item{epsilon}{an equivalance threshold}
}
\value{
a list of outputs.
}
\description{
Conduct an equivalence test of q1(v) = q0(v)
}
