% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-functions.R
\name{fit_ord_probit}
\alias{fit_ord_probit}
\title{Probit Regression}
\usage{
fit_ord_probit(Y, init = NULL, cut)
}
\arguments{
\item{Y}{A vector of outcome}
}
\description{
Fitting ordered probit regression
}
\keyword{internal}
