% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-functions.R
\name{fit_ord_probit_gr}
\alias{fit_ord_probit_gr}
\title{Probit regression with group dummies}
\usage{
fit_ord_probit_gr(Y, id_group, init = NULL, cut)
}
\arguments{
\item{Y}{A vector of the ordinal outcome.}

\item{id_group}{A vector of group indicator.}
}
\description{
Probit regression with group dummies
}
\keyword{internal}
