% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-data.R
\docType{data}
\name{gun_threewave}
\alias{gun_threewave}
\title{Three-wave Panel from Barney and Schaffner (2018)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 23792 rows and 7 columns.}
\usage{
gun_threewave
}
\description{
Three-wave Panel from Barney and Schaffner (2018)
}
\references{
Barney, David J., and Brian F. Schaffner. 
 "Reexamining the Effect of Mass Shootings on Public Support for Gun Control." 
  British Journal of Political Science 49.4 (2019): 1555-1565. 
  \url{https://doi.org/10.1017/S0007123418000352}.
}
\keyword{dataset}
