% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-data.R
\docType{data}
\name{gun_twowave}
\alias{gun_twowave}
\title{Two-wave panel from Barney and Schaffner (2019)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33106 rows and 11 columns.
  \describe{
     \item{caseid}{Unique id of a respondent.}
     \item{year}{Year of the response.}
     \item{guns}{Attitudes towards gun control, corresponding to a variable `CC320' in CCES. 
                 (1) "Less Strict", (2) "Kept As They Are" and (3) "More Strict".}
     \item{pds_25mi}{``Prior exposure'' variable, taking (1) if the mass shootings happened within 25 miles between 2000 and 2010.}
     \item{pds_100mi}{``Prior exposure'' variable, taking (1) if the mass shootings happened within 100 miles between 2000 and 2010.}
     \item{reszip}{Zipcode of a respondent.}
     \item{party2010}{3-point scale party id constructed by Barney and Schaffner (2019) based on `pid7' in 2010.}
     \item{pid3}{3-point scale party id, measured in 2010.}
     \item{pid7}{7-point scale party id, measured in 2010.}
     \item{treat_25mi}{Treatment indicator which takes 1 if the mass shootings happened within 25 miles.}
     \item{treat_100mi}{Treatment indicator which takes 1 if the mass shootings happened within 100 miles.}
   }}
\source{
Barney, David; Schaffner, Brian. 
 ``Replication Data for: Reexamining the Effect of Mass Shootings on Public Support for Gun Control.''
 \url{https://doi.org/10.7910/DVN/YJQIXP}, Harvard Dataverse, V1, 2018.

 Ansolabehere, Stephen; Schaffner, Brian. 
  ``2010 - 2012 CCES Panel Study/''
  \url{https://doi.org/10.7910/DVN/24416}, Harvard Dataverse, V4, 2014.
}
\usage{
gun_twowave
}
\description{
\code{gun_twowave} is a part of the replication data from Barney and Schaffner (2019).
The data set is constructed combining the two-wave panel from the Cooperative Congressional Election Study (CCES) 2010-12 study 
 with the mass shooting data originally analyzed by Newman and Hartman (2019).
}
\examples{
 require(dplyr)
 require(haven)
 data("gun_twowave")
 gun_twowave
}
\references{
Barney, David J., and Brian F. Schaffner. 
 "Reexamining the Effect of Mass Shootings on Public Support for Gun Control." 
  British Journal of Political Science 49.4 (2019): 1555-1565. 
  \url{https://doi.org/10.1017/S0007123418000352}.

Newman, B. J. and Hartman, T. K.
 ``Mass shootings and public support for gun control.''
 British Journal of Political Science, 49.4 (2019): 1527–1553.
 \url{https://doi.org/10.1017/S0007123417000333}.
}
\keyword{dataset}
