% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-functions.R
\name{log_like_probit_group}
\alias{log_like_probit_group}
\title{Log-likelihood Function}
\usage{
log_like_probit_group(par, Y, cut, id_group)
}
\arguments{
\item{par}{A vector of parameters.
means: \code{par[1:3]}; scale: \code{par[4:6]}; cutoff: \code{par[7:]} (only when J > 3)}

\item{Y}{A vector of categorical outcome.}

\item{cut}{A set of two cutoffs fixed for identification. Default is \code{cut = c(0, 1)}.}

\item{id_group}{A vector of group indicator.}
}
\description{
Log-likelihood Function
}
\keyword{internal}
