% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-checks.R
\name{ord_did_check_input}
\alias{ord_did_check_input}
\title{Input check for \code{ord_did}.}
\usage{
ord_did_check_input(Ynew, Yold, treat, id_cluster, n_boot)
}
\arguments{
\item{Ynew}{Ynew from \code{ord_did()}.}

\item{Yold}{Yold from \code{ord_did()}.}

\item{treat}{treat from \code{ord_did()}.}

\item{id_cluster}{id_cluster from \code{ord_did()}.}

\item{n_boot}{n_boot from \code{ord_did()}.}
}
\description{
Input check for \code{ord_did}.
}
\keyword{internal}
