% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-functions.R
\name{ord_did_run}
\alias{ord_did_run}
\title{Parameter Estimation for the Counterfactual Outcome}
\usage{
ord_did_run(Ynew, Yold, treat, cut, pre = FALSE)
}
\description{
Parameter Estimation for the Counterfactual Outcome
}
\keyword{internal}
