% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-package.R
\docType{package}
\name{orddid-package}
\alias{orddid}
\alias{orddid-package}
\title{Difference-in-Differences for Ordinal Outcomes}
\description{
This package provides tools for estimating treatment effects
 and diagnosing assumptions with ordinal outcomes in the difference-in-differences design.
}
\details{
The package has the following main functions:
\itemize{
  \item \code{\link{ord_did}}: Implements the ordinal DID estimator.
  \item \code{\link{summary.orddid}}: Output the treatment effects and their uncertainties.
  \item \code{\link{equivalence_test}}: Test the assumption in the pre-treatment periods.
  \item \code{\link{plot.orddid.test}}: Visualize the result of the equivalance test.
}

In addition to the above functions, \code{orddid} packages contains two example datasets:
 \code{\link{gun_twowave}} and \code{\link{gun_threewave}}.
}
\author{
\strong{Maintainer}: Soichiro Yamauchi \email{syamauchi@g.harvard.edu} (0000-0003-0554-2717)

}
