% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid.R
\name{orddid}
\alias{orddid}
\title{Ordinal Difference-in-Differences}
\usage{
orddid(Ynew, Yold, treat, cut, n_boot = 500)
}
\arguments{
\item{Ynew}{a vector of outcome for the post-treatment period.}

\item{Yold}{a vector of outcome for the pre-treatment period.}

\item{treat}{a vector of treatment indicator.}

\item{cut}{a vector of cutoffs.}

\item{n_boot}{the number of boostrapt iterations for computing uncertainty. Default is 500.}
}
\value{
A list
}
\description{
A function to implement Ordinal Difference-in-Differences method.
}
