% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-plot.R
\name{plot.orddid.test}
\alias{plot.orddid.test}
\title{Visualizing the result of equivalence_test()}
\usage{
\method{plot}{orddid.test}(obj, ylim, fill = TRUE, ...)
}
\arguments{
\item{obj}{An output from \code{equivalence_test()}.}

\item{ylim}{Range of y-axis of the plot.}

\item{fill}{A boolean argument; if \code{TRUE}, confidence band is filled by a polygon.}

\item{...}{Other arguments passed to \code{plot()} function.}
}
\description{
\code{plot.orddid.test} generates a plot that visualize the result of \code{\link{equivalence_test}}.
}
