% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordidd-equivalence.r
\name{plot_equivalence}
\alias{plot_equivalence}
\title{Plot equivalence test result}
\usage{
plot_equivalence(equivalence_result)
}
\arguments{
\item{equivalence_result}{An output from \code{\link{run_equivalence}}.}
}
\description{
Plot equivalence test result
}
