% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordidd-equivalence.r
\name{run_equivalence}
\alias{run_equivalence}
\title{Run equivalence test}
\usage{
run_equivalence(
  data,
  outcome,
  post,
  treat,
  cluster,
  n_boot = 500,
  use_parallel = FALSE,
  v_range = seq(0, 1, by = 0.01)
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{outcome}{Character scalar; column name for the outcome variable.}

\item{post}{Character scalar; column name for the post-period indicator (logical).}

\item{treat}{Character scalar; column name for the treatment indicator (logical).}

\item{cluster}{Character scalar; column name for the cluster id.
If left as \code{NULL}, bootstrap is implemented at the individual level.}

\item{n_boot}{The number of boostrapt iterations for estimating the variance. Default is \code{n_boot = 500}.}

\item{use_parallel}{Logical; if TRUE, use parallel computing for bootstrap.}

\item{v_range}{Numeric vector; the range of quantiles to evaluate the equivalence test.}
}
\description{
Run equivalence test
}
