% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-summary.R
\name{summarise_equivalence_test}
\alias{summarise_equivalence_test}
\title{Summarise function for equivalence_test()}
\usage{
summarise_equivalence_test(obj)
}
\description{
Summarise function for equivalence_test()
}
\keyword{internal}
