% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-summary.R
\name{summarise_ord_did}
\alias{summarise_ord_did}
\title{Summarise function for ord_did()}
\usage{
summarise_ord_did(obj, cumulative)
}
\description{
Summarise function for ord_did()
}
\keyword{internal}
