% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-diagnostic.R
\name{tv_gradient}
\alias{tv_gradient}
\title{Compute gradient of t(v; theta)}
\usage{
tv_gradient(v, theta)
}
\arguments{
\item{theta}{a list of parameters}
}
\description{
Compute gradient of t(v; theta)
}
