% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-diagnostic.R
\name{wald_test}
\alias{wald_test}
\title{Wald-based Falsification test for the distributional parallel trends assumption}
\usage{
wald_test(object)
}
\arguments{
\item{object}{An output from \code{orddid} function.}
}
\description{
Wald-based Falsification test for the distributional parallel trends assumption
}
