% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-diagnostic.R
\name{calc_threshold}
\alias{calc_threshold}
\title{Selecting Equivalence Threshold}
\usage{
calc_threshold(object, omega = 0.05)
}
\arguments{
\item{object}{a fitted object from \code{orddid}.}
}
\value{
a scalar value of threshold.
}
\description{
A function to compute the data-dependent threshold for the equivalence test.
}
