% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-probit.R
\name{.FitOrderedProbit}
\alias{.FitOrderedProbit}
\title{Fit Probit Regression}
\usage{
.FitOrderedProbit(Y, init = NULL, cutoff_val = NULL)
}
\arguments{
\item{Y}{A vector of outcome}
}
\description{
Fitting ordered probit regression
}
\keyword{internal}
