% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-diagnostic.R
\name{equivalence_test}
\alias{equivalence_test}
\title{Conduct an equivalence test of q1(v) = q0(v)}
\usage{
equivalence_test(object, alpha = 0.05, threshold = NULL)
}
\arguments{
\item{object}{a fitted object from \code{orddid}.}

\item{alpha}{level of a test. Should be between 0 and 1. Default is 0.05.}

\item{threshold}{an equivalance threshold.}
}
\value{
a list of outputs.
}
\description{
Conduct an equivalence test of q1(v) = q0(v)
}
