% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid.R
\name{ord_did}
\alias{ord_did}
\title{Ordinal Difference-in-Differences for Panel Data}
\usage{
ord_did(Ynew, Yold, treat, id_cluster = NULL, cut = c(0, 1), n_boot,
  pre = FALSE, verbose = FALSE)
}
\arguments{
\item{Ynew}{a vector of outcome for the post-treatment period.}

\item{Yold}{a vector of outcome for the pre-treatment period.}

\item{treat}{a vector of treatment indicator.}

\item{id_cluster}{a vector of cluster id.
If left \code{NULL}, bootstrap is blocked at the individaul level}

\item{cut}{a vector of cutoffs.}

\item{n_boot}{the number of boostrapt iterations for computing uncertainty. Default is 500.}

\item{pre}{a boolean to indicate if data comes from entirely pre-treatment periods.}

\item{verbose}{If \code{TRUE}, print the current iteration in the bootstrap.}
}
\value{
A list
}
\description{
A function to implement Ordinal Difference-in-Differences method.
}
