% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid.R
\name{ord_did}
\alias{ord_did}
\title{Ordinal Difference-in-Differences for Panel Data}
\usage{
ord_did(
  data,
  outcome,
  post,
  treat,
  cluster,
  n_boot = 500,
  use_parallel = FALSE
)
}
\arguments{
\item{treat}{A numeric vector of treatment indicator.
The treatment group should take 1 and the control group should take 0.}

\item{n_boot}{The number of boostrapt iterations for estimating the variance. Default is \code{n_boot = 500}.}

\item{Ynew}{A numeric vector of ordinal outcome for the post-treatment period.}

\item{Yold}{A numeric vector of ordinal outcome for the pre-treatment period.}

\item{id_cluster}{A vector of cluster id.
If left as \code{NULL}, bootstrap is implemented at the individual level.}

\item{cut}{A vector of cutoffs. Two numeric values should be specified. Default is \code{cut = c(0, 1)}.}

\item{pre}{A boolean argument used to indicate if the data comes entirely from pre-treatment periods.
This should be \code{TRUE} when the output is supplied to \code{\link{equivalence_test}}.}

\item{verbose}{If \code{TRUE}, print the progress of bootstrap iterations.}
}
\value{
\code{ord_did()} returns a list of class `orddid' containing the following components:
\item{fit}{A list with the output of the ordinal DID estimators,
           which contains parameter estimates and predicted probabilities for each category.}
\item{boot}{A list with the output of bootstraps,
            which contains parameter estimates and predicted probabilities for each category.}
\item{boot_params}{A list with all objects generated during the bootstrap step.}
}
\description{
\code{ord_did()} implements the difference-in-differences for the ordinal outcome.
}
\examples{
\donttest{
## load packages
library(orddid)
library(dplyr)

## load example data
data("gun_twowave")

## run
## fit the ordinal DID
set.seed(1234)
treated_id <- gun_twowave \%>\%
  filter(treat_25mi == 1) \%>\%
  select(caseid, treated_group = treat_25mi)
dat_use <- gun_twowave \%>\%
  left_join(treated_id, by = "caseid") \%>\%
  mutate(
    post = ifelse(year == 2012, TRUE, FALSE),
    treated = ifelse(!is.na(treated_group), TRUE, FALSE)
  )

dat_use \%>\%
  group_by(pid3) \%>\%
  group_map(
    ~ ord_did(
      data = .x,
      outcome = "guns",
      post = "post",
      treat = "treated",
      cluster = "caseid"
    )
  )

## view summary of the output
## non-cumulative effects
summary(fit, cumulative = FALSE)

## cumulative effects
summary(fit)
}
}
