% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-functions.R
\name{ord_did_boot}
\alias{ord_did_boot}
\title{Bootstrap function}
\usage{
ord_did_boot(Ynew, Yold, treat, cut, id_cluster, J, n_boot, verbose)
}
\description{
Conduct a block bootstrap at the unit level to compute variances and CIs
}
\keyword{internal}
