% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-checks.R
\name{ord_did_check_input}
\alias{ord_did_check_input}
\title{Input check for \code{ord_did_run}.}
\usage{
ord_did_check_input(Ynew, Yold, treat)
}
\arguments{
\item{Ynew}{Ynew from \code{ord_did_run()}.}

\item{Yold}{Yold from \code{ord_did_run()}.}

\item{treat}{treat from \code{ord_did_run()}.}
}
\description{
Input check for \code{ord_did_run}.
}
\keyword{internal}
