% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orddid-summary.R
\name{summary.orddid}
\alias{summary.orddid}
\title{Summary function}
\usage{
\method{summary}{orddid}(obj, cumulative = TRUE)
}
\arguments{
\item{obj}{an output object from \link{\code{orddid}} or \link{\code{equivalence_test}}.}

\item{cumulative}{Default is \code{TRUE}. Only effective for displaying the causal effects.
If \code{TRUE}, cumulative effect is presented.}
}
\description{
Summary function
}
